requirements.txt
asgiref==3.8.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-cuser==2017.3.16
django-extensions==3.1.0
django-modeltranslation==0.16
django-session-timeout==0.1.0
django-tinymce==4.1.0
idna==2.10
pillow==11.2.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
typing_extensions==4.13.2
urllib3==1.26.3
